requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
werkzeug>=2.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
replicate>=0.22.0